requirements.txt
faiss-cpu
sentence-transformers
requests
